Переносит тест на клик в новый компонент small-movie-card.e2e.test.js
